First draft of adding a check-and-set to deleteKV
The main dilemma with adding check-and-set to deleteKV is that all the
delete KV methods return Response<Void>. Consequently we can't actually
return whether the c-a-s worked. We can't change Response<Void> to
Response<Boolean> without breaking a lot of existing code.

This just adds the capability for check-and-set but punts on the
Response<Void> problem. If callers want to see if the delete succeeded,
they'll have to do another getKV. However, this is the least invasive
version of this change.

https://github.com/Ecwid/consul-api/issues/49