fix: #186 token as request param is deprecated
Consul now warns aggressively when token is provided as a request param
and the X-Consul-Token is required. Also as already mentioned in #186
token request param will be removed in Consul v1.17.

By reading the code and related PRs I've seen that in the so called
"new http architecture" this is addressed although not all clients fully
utilize the new Request class.

This PR uses the Request class for both Catalog and KV clients.

Should address both #186 and #237.